Add files via upload
This project is made for the company "Gymso fitness" which runs multiple fitness centers throughout India. Each gym offers a variety of cardio and strength training equipment as well as group exercise programs. 
The focus of this website design is to grow awareness of the brand and to establish and maintain the customer relationship through the member portal where members can get easy access to important information. The members can register and log in to the website. Website design elements like colors, fonts, images, etc. all bring about the brand identity. Also, the website’s layout is very simple, neat, and designed in a matter that will bring much attention to the most essential parts.
The project has been implemented based on the front-end and back-end technologies learned. The front-end is created using HTML, CSS, JavaScript, and Angular JS for making the website dynamic. For the backend node, JS is used and for the database, both Mongo DB and MySQL are used.